# star-sdk-backend/starsdk-ws/generated/swagger/swagger.yaml
openapi: 3.0.0
servers:
- url: https://app-test-ws.star-app.ch
  description: ''
- url: https://app-prod-ws.star-app.ch
  description: ''
info:
  version: 1.0-develop
  description: STARSDK API
  title: STARSDK API
paths:
  /v1/:
    get:
      summary: hello
      description: hello
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
  /v1/exposed:
    post:
      summary: addExposee
      description: Enpoint used to publish the SecretKey.
      responses:
        '200':
          description: Returns OK if successful
          content:
            application/json:
              schema:
                type: string
        '400':
          description: "Multiple possible errors: \n- Key is not base64\n - No AuthData\
            \ given"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ch.ubique.starsdk.model.ExposeeRequest'
        description: The ExposeeRequest contains the SecretKey from the guessed infection
          date, the infection date itself, and some authentication data to verify
          the test result
      parameters:
      - name: User-Agent
        in: header
        description: App Identifier (PackageName/BundleIdentifier) + App-Version +
          OS (Android/iOS) + OS-Version
        required: true
        schema:
          type: string
  /v1/exposed/{dayDateStr}:
    get:
      summary: getExposed
      description: ''
      responses:
        '200':
          description: Returns ExposedOverview, which includes all secretkeys which
            were published on _dayDateStr_.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ch.ubique.starsdk.model.ExposedOverview'
        '500':
          description: If the date time string is not parseable.
      parameters:
      - name: dayDateStr
        in: path
        description: The date for which we want to get the SecretKey.
        required: true
        schema:
          type: string
components:
  schemas:
    ch.ubique.starsdk.model.AuthDataMethod:
      type: string
      enum:
      - NONE
      - REDEEM_CODE
    ch.ubique.starsdk.model.ExposedOverview:
      type: object
      properties:
        exposed:
          type: array
          items:
            $ref: '#/components/schemas/ch.ubique.starsdk.model.Exposee'
          description: A list of all SecretKeys
          example: ''
    ch.ubique.starsdk.model.Exposee:
      type: object
      required:
      - key
      - onset
      properties:
        Id:
          type: integer
        key:
          type: string
          description: The SecretKey of an exposed.
          example: AAAA
        onset:
          type: string
          description: The onset of an exposed.
          example: '2020-04-06'
    ch.ubique.starsdk.model.ExposeeAuthData:
      type: object
      properties:
        value:
          type: string
          description: Authentication data used to verify the test result
          example: ''
        method:
          $ref: '#/components/schemas/ch.ubique.starsdk.model.AuthDataMethod'
          description: Specify the method, which should be used to verify data
          example: ''
    ch.ubique.starsdk.model.ExposeeRequest:
      type: object
      required:
      - key
      - onset
      - authData
      properties:
        key:
          type: string
          description: The SecretKey used to generate EphID
          example: ''
        onset:
          type: string
          description: 'The onset date of the secret key. Format: yyyy-MM-dd'
          example: ''
        authData:
          $ref: '#/components/schemas/ch.ubique.starsdk.model.ExposeeAuthData'
          description: AuthenticationData provided by the health institutes to verify
            the test results
          example: ''